Refactor vitessce config
